refactor: MainPage images move to assets
