+++ src/pages/main_page/ApplyCardMain/MainApplyCard.css
