+++ src/pages/main_page/ImgMain/ImgMain_1.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import img_1 from "../components/img/main_img_1.png";
+import img_1 from "../../../assets/images/main/main_img_1.png";
 
 import GoToApply from "../components/GoToApply";
