+++ src/pages/main_page/Main.js
@@ -3,7 +3,7 @@ import React from "react";
 // components
 import ImgMain from "./ImgMain/ImgMain";
 import IntroduceMain from "./IntroduceMain/IntroduceMain";
-import MainApplyCard from "./IntroduceMain/ApplyCardMain/MainApplyCard";
+import MainApplyCard from "./ApplyCardMain/MainApplyCard";
 
 const Main = () => {
   return (
